refactor(validate): split 870-line god-file into a package — H4 (audit)
The validate module was approaching 900 lines: 9 independent visual
checks, plus shared types and an orchestrator, in a single .py. Each
check was 60–120 lines with nested loops and bbox geometry, so the
file was painful to navigate and the checks couldn't be reasoned
about in isolation.

This commit reshapes the module without changing the public surface:

    src/dartwork_mpl/validate/
    ├── __init__.py          — re-exports the public surface
    ├── _types.py            — Severity, VisualWarning, BBOX_ERRORS
    ├── _orchestrator.py     — validate_figure
    └── _checks/
        ├── __init__.py
        ├── overflow.py          (was _check_overflow)
        ├── overlap.py           (was _check_overlap)
        ├── legend.py            (was _check_legend_overflow)
        ├── tick_crowd.py        (was _check_tick_crowding + helper)
        ├── empty_axes.py        (was _check_empty_axes)
        ├── margin.py            (was _check_margin_asymmetry)
        ├── pie_label.py         (was _check_pie_label_offset)
        ├── clipped_text.py      (was _check_clipped_text)
        └── cross_axes.py        (was _check_cross_axes_overlap)

Public API
----------
Unchanged. `from dartwork_mpl.validate import Severity, VisualWarning,
validate_figure` still works; `dm.validate_figure(fig)` still works.

Renames
-------
The internal helper functions lost their `_check_` prefix because
they're now module-private under `_checks/`. `_BBOX_ERRORS` became
`BBOX_ERRORS` (still package-private via `_types`). Constants moved
next to the check that uses them.

Each module is now ≤200 lines.

Verification
------------
- Public-surface smoke: `dm.validate_figure(fig)` → identical output
- `pytest tests/test_validate.py tests/test_validate_fixes.py
  tests/test_preset_matrix.py tests/test_layout.py tests/test_io.py`
  → 120 passed.
- `ruff check src/dartwork_mpl/validate/` → All checks passed.
- `mypy src/dartwork_mpl/validate/` → Success: no issues found in 13 files.

Paired with the simple_layout step extraction (H5) in the next commit.
