+++ src/dartwork_mpl/validate/_checks/legend.py
@@ -1,0 +1,63 @@
+"""LEGEND_OVERFLOW: legend that occupies too much of its axes area."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+from .._types import BBOX_ERRORS, Severity, VisualWarning
+
+if TYPE_CHECKING:
+    from matplotlib.backend_bases import RendererBase
+    from matplotlib.figure import Figure
+
+__all__ = ["check_legend_overflow"]
+
+_THRESHOLD = 0.30  # 30% of axes area
+
+
+def check_legend_overflow(
+    fig: Figure, renderer: RendererBase
+) -> list[VisualWarning]:
+    """Detect legends consuming too large a fraction of the Axes area."""
+    warnings: list[VisualWarning] = []
+
+    for i, ax in enumerate(fig.axes):
+        legend = ax.get_legend()
+        if legend is None or not legend.get_visible():
+            continue
+        try:
+            leg_ext = legend.get_window_extent(renderer)
+            ax_ext = ax.get_window_extent(renderer)
+        except BBOX_ERRORS:
+            continue
+
+        ax_area = ax_ext.width * ax_ext.height
+        if ax_area <= 0:
+            continue
+
+        # Intersection of legend bbox with axes bbox
+        x0 = max(leg_ext.x0, ax_ext.x0)
+        y0 = max(leg_ext.y0, ax_ext.y0)
+        x1 = min(leg_ext.x1, ax_ext.x1)
+        y1 = min(leg_ext.y1, ax_ext.y1)
+        overlap_area = max(0, x1 - x0) * max(0, y1 - y0)
+        ratio = overlap_area / ax_area
+
+        if ratio > _THRESHOLD:
+            warnings.append(
+                VisualWarning(
+                    severity=Severity.WARNING,
+                    check_id="LEGEND_OVERFLOW",
+                    message=(
+                        f"Legend occupies {ratio:.1%} of axes[{i}] area "
+                        f"(threshold: {_THRESHOLD:.0%})"
+                    ),
+                    detail={
+                        "axes_index": i,
+                        "ratio": round(ratio, 3),
+                        "threshold": _THRESHOLD,
+                    },
+                )
+            )
+
+    return warnings
